Move team management and brag on slack scipts to 'management_utils' for clarity
