+++ core/management/commands/copy_event.py
@@ -5,7 +5,7 @@ import djclick as click
 
 from core.command_helpers import gather_event_date_from_prompt
 from core.models import Event
-from core.utils import get_main_organizer, get_team, create_users, brag_on_slack_bang
+from core.management_utils import get_main_organizer, get_team, create_users, brag_on_slack_bang
 
 
 def get_event(id_str):
